🚧 WIP connection/disconnection

+++ graphql/RootSchema.js
@@ -239,27 +239,31 @@ class UnauthorizedError extends Error {
   }
 }
 
+const isAuth = resolver => (obj, args = {}, context, info) => {
+  if(!context.user) throw new UnauthorizedError()
+  resolver(obj, args = {}, context, info)
+}
+
 const resolvers = {
   MAP: GraphQLJSON,
   Query: {
-    book: (obj, args = {}, context, info) => {
+    book: isAuth((obj, args = {}, context, info) => {
       const { id } = args
       const filters = {}
       if (id) filters._id = id
       return Book.findOne(filters)
-    },
+    }),
     books: (obj, args = {}, context, info) => {
-      if(!context.user) throw new UnauthorizedError()
       const { author, draft } = args
       const filters = {}
       if (author) filters.authorId = author
       if (typeof draft === 'boolean') filters.draft = draft
       return Book.find(filters)
     },
-    author: (obj, { id }, context, info) => User.findById(id),
-    page: (obj, { bookId, pageId }, context, info) => Book.findById(bookId).then(book => book.pages.id(pageId)),
+    author: isAuth((obj, { id }, context, info) => User.findById(id)),
+    page: isAuth((obj, { bookId, pageId }, context, info) => Book.findById(bookId).then(book => book.pages.id(pageId))),
 
-    tryGame: (_, { bookId, playerId }) => Book.findById(bookId).then(book => {
+    tryGame: isAuth((_, { bookId, playerId }) => Book.findById(bookId).then(book => {
       const stats = book.stats.reduce((acc, stat) => {
         return {
           ...acc,
@@ -268,7 +272,6 @@ const resolvers = {
       }, {})
 
       const objects = book.objects.reduce((acc, object) => {
-        // console.log(object.atStart)
         if (object.atStart) return [...acc, object._id]
         return acc
       }, [])
@@ -280,7 +283,7 @@ const resolvers = {
         stats,
         objects,
       })
-    }),
+    })),
   },
   Book: {
     author: (book) => {
@@ -288,58 +291,57 @@ const resolvers = {
     }
   },
   User: {
-    drafts: (user) => {
+    drafts: isAuth((user) => {
       return Book.find({ authorId: user.id, draft: true })
-    },
-    publications: (user) => {
+    }),
+    publications: isAuth((user) => {
       return Book.find({ authorId: user.id, draft: false })
-    },
+    }),
   },
   Mutation: {
-    createBook: (_, { author }) => {
+    createBook: isAuth((_, { author }) => {
       const book = new Book({
         authorId: author,
       })
       return book.save().then(book => ({ id: book.authorId }))
-    },
+    }),
     // Book
-    updateBook: (_, { book }) => Book.findByIdAndUpdate(book.id, book, { new: true }),
-    deleteBook: (_, { id }) => Book.findByIdAndRemove(id).then(book => ({ id: book.authorId })),
-    publishBook: (_, { id }) => Book.findByIdAndUpdate(id, { draft: false }, { new: true }).then(book => ({ id: book.authorId })),
-    unpublishBook: (_, { id }) => Book.findByIdAndUpdate(id, { draft: true }, { new: true }).then(book => ({ id: book.authorId })),
+    updateBook: isAuth((_, { book }) => Book.findByIdAndUpdate(book.id, book, { new: true })),
+    deleteBook: isAuth((_, { id }) => Book.findByIdAndRemove(id).then(book => ({ id: book.authorId }))),
+    publishBook: isAuth((_, { id }) => Book.findByIdAndUpdate(id, { draft: false }, { new: true }).then(book => ({ id: book.authorId }))),
+    unpublishBook: isAuth((_, { id }) => Book.findByIdAndUpdate(id, { draft: true }, { new: true }).then(book => ({ id: book.authorId }))),
 
     // createStat: (_, { bookId }) => createBookSubRessource('stats', bookId, new Stat()),
-    createStat: (_, { bookId }) => findBookById(bookId).then(book => {
+    createStat: isAuth((_, { bookId }) => findBookById(bookId).then(book => {
       return book.addOne('stats', new Stat())
                  .save()
-    }),
-    updateStat: (_, { bookId, stat }) => findBookById(bookId).then(book => {
+    })),
+    updateStat: isAuth((_, { bookId, stat }) => findBookById(bookId).then(book => {
       return book.selectOne('stats', stat.id)
                  .set(stat)
                  .save()
-    }),
-    deleteStat: (_, { bookId, statId }) => findBookById(bookId).then(book => {
+    })),
+    deleteStat: isAuth((_, { bookId, statId }) => findBookById(bookId).then(book => {
       return book.deleteOne('stats', statId)
                  .save()
-    }),
+    })),
 
-    createPageReturnPage: (_, { bookId }) => findBookById(bookId).then(book => {
+    createPageReturnPage: isAuth((_, { bookId }) => findBookById(bookId).then(book => {
       const page = new Page()
       return book.addOne('pages', page)
                  .save()
                  .then(() => page)
-    }),
-    createPage: (_, { bookId }) => findBookById(bookId).then(book => {
+    })),
+    createPage: isAuth((_, { bookId }) => findBookById(bookId).then(book => {
       return book.addOne('pages', new Page())
                  .save()
-    }),
-    updatePage: (_, { bookId, page }) => findBookById(bookId).then(book => {
+    })),
+    updatePage: isAuth((_, { bookId, page }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', page.id)
                  .set(page)
                  .save()
-    }),
-    deletePage: (_, { bookId, pageId }) => findBookById(bookId).then(book => {
-
+    })),
+    deletePage: isAuth((_, { bookId, pageId }) => findBookById(bookId).then(book => {
       book.ressource.pages.forEach(page => {
         page.transitions.forEach(transition => {
           if (transition.toPage && transition.toPage === pageId) {
@@ -349,97 +351,97 @@ const resolvers = {
       })
       return book.deleteOne('pages', pageId)
                  .save()
-    }),
+    })),
 
-    createObject: (_, { bookId }) => findBookById(bookId).then(book => {
+    createObject: isAuth((_, { bookId }) => findBookById(bookId).then(book => {
       return book.addOne('objects', new ObjectModel())
                  .save()
-    }),
-    updateObject: (_, { bookId, object }) => findBookById(bookId).then(book => {
+    })),
+    updateObject: isAuth((_, { bookId, object }) => findBookById(bookId).then(book => {
       return book.selectOne('objects', object.id)
                  .set(object)
                  .save()
-    }),
-    deleteObject: (_, { bookId, objectId }) => findBookById(bookId).then(book => {
+    })),
+    deleteObject: isAuth((_, { bookId, objectId }) => findBookById(bookId).then(book => {
       return book.deleteOne('objects', objectId)
                  .save()
-    }),
+    })),
 
-    createPageEffect: (_, { bookId, pageId }) => findBookById(bookId).then(book => {
+    createPageEffect: isAuth((_, { bookId, pageId }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .addOne('effects', new Effect())
                  .save()
-    }),
-    updatePageEffect: (_, { bookId, pageId, effect }) => findBookById(bookId).then(book => {
+    })),
+    updatePageEffect: isAuth((_, { bookId, pageId, effect }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .selectOne('effects', effect.id)
                  .set(effect)
                  .save()
-    }),
-    deletePageEffect: (_, { bookId, pageId, effectId }) => findBookById(bookId).then(book => {
+    })),
+    deletePageEffect: isAuth((_, { bookId, pageId, effectId }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .deleteOne('effects', effectId)
                  .save()
-    }),
+    })),
 
-    createPageTransition: (_, { bookId, pageId }) => findBookById(bookId).then(book => {
+    createPageTransition: isAuth((_, { bookId, pageId }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .addOne('transitions', new Transition({ fromPage: pageId }))
                  .save()
-    }),
-    updatePageTransition: (_, { bookId, pageId, transition }) => findBookById(bookId).then(book => {
+    })),
+    updatePageTransition: isAuth((_, { bookId, pageId, transition }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .selectOne('transitions', transition.id)
                  .set(transition)
                  .save()
-    }),
-    deletePageTransition: (_, { bookId, pageId, transitionId }) => findBookById(bookId).then(book => {
+    })),
+    deletePageTransition: isAuth((_, { bookId, pageId, transitionId }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .deleteOne('transitions', transitionId)
                  .save()
-    }),
+    })),
 
-    createPageTransitionCondition: (_, { bookId, pageId, transitionId }) => findBookById(bookId).then(book => {
+    createPageTransitionCondition: isAuth((_, { bookId, pageId, transitionId }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .selectOne('transitions', transitionId)
                  .addOne('conditions', new Effect())
                  .save()
-    }),
-    updatePageTransitionCondition: (_, { bookId, pageId, transitionId, condition }) => findBookById(bookId).then(book => {
+    })),
+    updatePageTransitionCondition: isAuth((_, { bookId, pageId, transitionId, condition }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .selectOne('transitions', transitionId)
                  .selectOne('conditions', condition.id)
                  .set(condition)
                  .save()
-    }),
-    deletePageTransitionCondition: (_, { bookId, pageId, transitionId, conditionId }) => findBookById(bookId).then(book => {
+    })),
+    deletePageTransitionCondition: isAuth((_, { bookId, pageId, transitionId, conditionId }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .selectOne('transitions', transitionId)
                  .deleteOne('conditions', conditionId)
                  .save()
-    }),
+    })),
 
-    createPageTransitionEffect: (_, { bookId, pageId, transitionId }) => findBookById(bookId).then(book => {
+    createPageTransitionEffect: isAuth((_, { bookId, pageId, transitionId }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .selectOne('transitions', transitionId)
                  .addOne('effects', new Effect())
                  .save()
-    }),
-    updatePageTransitionEffect: (_, { bookId, pageId, transitionId, effect }) => findBookById(bookId).then(book => {
+    })),
+    updatePageTransitionEffect: isAuth((_, { bookId, pageId, transitionId, effect }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .selectOne('transitions', transitionId)
                  .selectOne('effects', effect.id)
                  .set(effect)
                  .save()
-    }),
-    deletePageTransitionEffect: (_, { bookId, pageId, transitionId, effectId }) => findBookById(bookId).then(book => {
+    })),
+    deletePageTransitionEffect: isAuth((_, { bookId, pageId, transitionId, effectId }) => findBookById(bookId).then(book => {
       return book.selectOne('pages', pageId)
                  .selectOne('transitions', transitionId)
                  .deleteOne('effects', effectId)
                  .save()
-    }),
+    })),
 
-    createGame: (_, { bookId, playerId }) => {
+    createGame: isAuth((_, { bookId, playerId }) => {
       // TODO: implement this (same as tryGame Query, but saving in Database)
       // const game = new Game({
       //   playerId,
@@ -449,9 +451,9 @@ const resolvers = {
       //   bookStatus:'up-to-date',
       //   stats,
       // }).then(game => game)
-    },
+    }),
 
-    updatePassword: async (_, { userId, oldPassword, newPassword, confirmation }) => {
+    updatePassword: isAuth((_, { userId, oldPassword, newPassword, confirmation }) => {
       const paramsNotEmpty = oldPassword === '' || newPassword === '' || confirmation === ''
       const passwordMatch = newPassword !== confirmation
 
@@ -471,7 +473,7 @@ const resolvers = {
         delete user.password
         return user
       })
-    },
+    }),
   }
 }
